Fix convergence issue when solving for time
In the unconstrained optimization problem, we were doing an undamped
step.  This was resulting in us getting stuck in cycles for some
problems.

The correct answer theoretically is to add line search.  So, add it.
Section 10.3 in https://web.stanford.edu/~boyd/cvxbook/bv_cvxbook.pdf
has the algorithm documented.

This takes a log which used to fail and makes it pass.

Change-Id: I79a7bc8e81aafefdcde5984e8f61cb253f6b0e71
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>