Make line search in our timestamp solving code more correct
We had a mix of the newton's method and equality constrained primal-dual
solver strategy in the same algorithm.  Move over to using the
primal-dual algorithm in
https://web.stanford.edu/~boyd/cvxbook/bv_cvxbook.pdf Algoritym 10.2
fully.

Line search was occasionally getting stuck at a local minima.  To fix
that, we add a limit to how small a step it will take.  This makes it so
we always make some progress, even if it isn't far, and solves both the
logs where we need line search, and the logs where line search hurts
convergence.

Change-Id: I464e3f65f79bd42c50f7bc52d00337d7b22bf515
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>