Make Rust FlatbufferBuilder realtime
We do this by changing the internal vectors used to ArrayVec
(https://docs.rs/arrayvec/latest/arrayvec/) which are effectively
a direct replacement to a Vec but backed by a statically
defined buffer, similar to an array but with variable size (note that
it's the capacity that is constant).

This is most likely just a workaround. Having looked at the C++
implementation of flatbuffer, it seems like they use the same
Allocator API to store this metadata as well. We probably will
want to do this in Rust as well at some point. For now we just
set hardlimits on the arrays.

Change-Id: I2d3d7e0e85d3bced25f0fd5604e776f311630ad7
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>