Make EventLoopRuntime a thin wrapper over &CppEventLoopRuntime
The idea with this is to (in a future CR), pass EventLoopRuntime
as an owned type instead of a reference. This is important
because we can move the CppEventLoopRuntime as part of the
EventLoopRuntimeHolder. Without that, it's pretty much impossible
to create an event loop task that uses the runtime internally
(e.g. to get the time).

Change-Id: I163d94edba8a57191040fef171bb7abad624f7f8
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>