Flip on --bfbs-builtins
This makes it so that "builtin" flatbuffer attributes (e.g.,
force_align) get included in the generated reflection data.
Because I want to be able to support force_align in particular
in the static flatbuffers API, we need the force_align data
to get included in the generated schema information.

Change-Id: Ife408e214b139f53ad3ab34a07551747b08646f7
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>