Fix an msan violation in AOS controls code
In a move, the destination can be uninitialized prior to the move (and
the origin should be considered uninitialized after the move).

A swap however assumes that both are initialized before (and thus after
as well).

The ::std::swap was ok with msan in clang 17, but the Eigen-swap was
not.

This is in preparation for updating to clang 17.

Change-Id: I2ef80f5f5a3b34ba8e8d2529ca5ff6c9436ffcd1
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>