Generate rules for linking against system libraries in the rootfs
This uses dpkg-query, objdump, and pkgconf to generate the list of files
in each package, the package dependency list, the list of shared object
files needed to link each package, and then link them all together.

The end result is that you can include "opencv4", and it does the right
thing with very minimal hand annotations in the python code.

I need to clean the script up a bit, add some comments, and split it out
so it can be shared with x86 so we can do this there.

The driving force here is that we need to upgrade all the bullseye
packages to bookworm to get them to run on the Orin...

Change-Id: I65e8c59acc6fa81e20b0c130e3fa0d144ecc5dda
Signed-off-by: Austin Schuh <austin.linux@gmail.com>