Mark our SCTP traffic to have higher IP precedence
We want to ensure that our SCTP traffic is unaffected by anything else
going on in the network.

We set the Type-Of-Service (TOS) to 176 at Chong Ding's suggestion:

> 176 happened to be the only one that ubuntu decided to let go into
> bucket 0. Also https://www.tucny.com/Home/dscp-tos

With this patch, I now see the packets being marked with `tos 0xb2`.
As far as I can tell, the 2 least significant bits there are not
relevant to our purposes. 0xb0 translates to 176 which is what we
want. The 2 least significant bits are the "Explicit Congestion
Notification" bits according to Wireshark.

Change-Id: I5726009edf408e96508aed813a167895269bbc53
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>