Actually detect apriltags
This does the line fits for the borders, finds the best up to 10
candidate corners, and picks the best one.  It then does the final
filtering to pick which quads to evaluate, evaluates them, and then uses
aprilrobotics to refine and report out the solution.

There's a small (1e-3) discrepency on some of the tags to track down,
but overall it works pretty well.  Takes 4ms to process the imx296
images.

Change-Id: I6cff5654252897f46def422c021b1dce74036476
Signed-off-by: Austin Schuh <austin.linux@gmail.com>