Reduce flakeness of shm_event_loop_test
On some platforms, we were seeing one of the send too fast tests fail
occasionally.  Fix it by giving the scheduler a bit more of a grace
period by sending batches of messages and waking up less often, and
running it on the RT scheduler to reduce impact of other load.

Change-Id: I940dc58e81fed0e7ab8ef4da6b089226bfbc9945
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>