Update Cypress and Chromium
I was debugging some QR code detection and wondered if the old
versions of these 2 pieces had anything to do with the issues.
Upgrading didn't help, but I figured it's worth keeping.

This patch updates Chromium to latest and Cypress to near-latest. I
had to pull in an upstream patch for rules_cypress because upstream
Cypress changed their released tarballs. So the checksums needed to
change. I didn't update Cypress to latest because of a reported issue:
https://github.com/aspect-build/rules_cypress/pull/56#issuecomment-1827125915

As far as I can tell, there are no functional changes.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I04592e367cf1de94fe4243c7542d06ce7b16a418