rename variables in aos Logger
Rename some variables to avoid confusion. In the logger there
is a configuration supplied to the constuctor, and another
configuration from the event_loop. When referring to channels,
be specific about which configuration we are referencing.

Rename channel_index to logged_channel_index.
Rename 'channel' to event_loop_channel.
Rename 'config_channel' to logged_channel.

Change-Id: Ib30081583b639e5df8779b116d27d11f9b83b070
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>