Add a localizer logger that doesn't log images
Add this so that we can have a logger running on the localizer pi which
does not log images and so can be logging 100% of the time.

We may want to actually log this into an entirely different directory
than the image logs so that if the image logs fill up the disk we still
get regular logs.

Change-Id: Ida4906ecfcc65604c01e5f7fd2f1e8773d3ed25a
Signed-off-by: James Kuszmaul <jabukuszmaul+collab@gmail.com>