Fix network dropouts by increasing buffer sizes.
We were only allocating 1k of buffer space since the max message size
was so small.  This was causing any backups in sending to immediately
result in send failures on the server side.

Force the send queue to 128k, and the receive queue to 2 MB to give us
enough space to make sure the network is never our problem.

Change-Id: Iff48089537221d3ee891caf641142f788f7deb81
Signed-off-by: Austin Schuh <austin.linux@gmail.com>