Convert non-pwm synchronized sensor reading to be a phased loop
We use this mode when we are synced to CAN.  The synchronous CAN stack
that CTRE provides is aligned with the monotonic clock.  So, a phased
loop is a better fit for what we are trying to do.

Old robots had no requirements here, so it is safe to just change them
all.

Change-Id: I77e3351f53c1c5ea9fe37eba09b7ffdaa46ee12e
Signed-off-by: Austin Schuh <austin.linux@gmail.com>