Logger: Define contract to record transmit time for reliable messages
The sent time for reliable messages are not very useful to estimate
network latency across reboots. Add a new field which provides
information about the transmit time, i.e time at which the message
was handed to the kernel to be sent over the network.

This can then be used as a valid timestamp point for the timestamp
solver particularly for reliable messages. It also gives us higher
visibility into the path of the message through the network.

Change-Id: I7f2ea6b0300894d542a42de9caddc57efa965253
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>