Add tests for individual node loggers to config_validators
This makes it so that you can easily test that each individual node's
logs can be used to replay all of the state on that log. While we
already theoretically had logic to exercise this in the config
validator, said logic was both incorrectly written and not actually
being used by any of the config validators.

Simultaneously: While upstreaming this, start reducing the number of
explicit @repo_name that we have written out in the repository.

Change-Id: I0c7cb634b3d7bbc63503c607a5f8d91bc184d416
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>