scouting: Re-authorize users when their auth expires
A few students noted at the San Francisco regional that they get 401
errors when trying to submit scouting data. I'm not entirely convinced
how to fix that, but this patch attempts to implement a work around.

Whenever a user tries to submit data, we check for the resulting
status code. If the code is 401, then the app now opens a new tab in
an attempt to re-authorize. The tab will close itself. When the tab
closes, the app will attempt to submit the scouting data again.

Change-Id: I3a4273c9c0f571f68332877f744c9c7013a7e66a
Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>