Add a test reproducing a log reading failure
We saw a case where we only had a reliable message for the second boot.

on pi2 from pi1 { "name": "/atest1", "type": "aos.examples.Ping" }
    .monotonic_event_time={.boot=0, .time=99.000150000sec},
    .monotonic_remote_time={.boot=0, .time=99.000000000sec},
    .monotonic_timestamp_time={.boot=0, .time=99.000250000sec}

on pi2 from pi1 { "name": "/atest2", "type": "aos.examples.Ping" }
    .monotonic_event_time={.boot=1, .time=0.000100000sec},
    .monotonic_remote_time={.boot=0, .time=1.000000000sec},
    .monotonic_timestamp_time={.boot=0, .time=100.000200000sec},

This ends up resulting in it looking like the second boot happened 50
seconds before the end of the first boot, and time going backwards.

Add a test to recreate this situation so we can fix it.

Change-Id: I4615e9c85ba57dcd7e656af1f94d3ca90f2c2126
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>