aos: Fix `--force_timestamp_loading` for single-node logs
We're currently seeing a bunch of segfaults in log-reading code.
This is happening because the timestamp loading code
performs some checks against a list of source nodes. This list was
only populated for multi-node logs.

The fix here is to populate the list (with zeroes) even for
single-node logs. That lets the code proceed happily.

Before this patch I got the following crash for the new test:

    $ bazel test --config=strict_ubsan @aos//aos/events/logging:logfile_utils_test -c opt
    ...
    [ RUN      ] SingleNodeTimestampMapperTest.QueueTimestampsForSingleNodes
    external/llvm_k8/include/c++/v1/vector:1548:12: runtime error: reference binding to null pointer of type 'const value_type' (aka 'const unsigned long')
        #0 0x7f404ba4197e  (/data/phil_cache/bazel/_bazel_philipp/d441cd67cd5f527804d5d757edac2e45/execroot/shasta/bazel-out/k8-opt-ubsan/bin/external/aos/aos/events/logging/../../../../../_solib_k8/libexternal_Saos_Saos_Sevents_Slogging_Sliblogfile_Uutils.so+0x24197e) (BuildId: 64e54686062e7b4cd33e15bb59eb809a)

References: PRO-26639
Change-Id: I5a28f09af4f77da9b9aa2f36030394bd6d7e2e77
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>