Fix transmit timestamp logic in SimulatedNetworkBridge
We were seeing a case when monotonic_remote_transmit_times_ got out of
sync with the fetcher.  This happened when another spurrious wakeup
triggered us to FetchNext out of sync with the actual wakeups.

This really just points to the logic being overly convoluted.  Use
monotonic_remote_transmit_times_ as the source of truth for everything
and sync the fetcher to it instead, and only feed it with timestamps on
the actual wakeups (reliable messages at startup + reconnect need some
special handling, but everything else is easy).

Change-Id: I9c1cabda98ea9765129e6b5347d2a676536d3872
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>