aos: Support generated replacements in jinja2_template
I have a rule that generates replacements for a Jinja2 template. I
can't use any of the current mechanisms to do this so I decided to
enhance the rule.

You can now specify a `parameters_file` attribute whose contents will
be added to the available variables inside the template.

This came up when I wanted to copy the spyglass colors into a foxglove
script.

I couldn't figure out a good way to write a build test for this
because the //build_tests package in BRT's vendored copy of AOS
doesn't load properly.

Change-Id: Ib046817e5b698d5358b787e67953503ed2b57f78
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>