Only schedule OnRun callbacks right after construction in aos
OnRun is meant to be a place to do work right before the event loop
starts to run.  On ShmEventLoop, this is the case, and restarting the
loop isn't all that well supported, so this doesn't make sense there
either.  A lot of our tests have used this behavior since it was easy
and not disallowed.

A follow-up commit will explicitly dis-allow this behavior.

Change-Id: If50dcc0e737725129ef36cb27e07be774d96c691
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>