scouting: Add more debug logging when stats get submitted
We're getting reports of folks not being able to submit data scouting
because of duplication errors. Looking at the database, there are some
actions present for a team's match, but no corresponding stats.
Subsequent attempts to submit the data will fail because the actions
are already there.

What we need to figure out is what error is happening the first time
around. I.e. what happens on the initial submission attempt? This
patch aims to help us find out by logging more information.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Iae7c3bd5237b434126cc55c04c786d69c9d76c6a