scouting: Lower the default QR code density to 450
The scouts are having trouble scanning the QR codes. They could
manually change the setting every time they want to save a QR code,
but it's easier to change the default for now.

Ideally, we'd make it so the setting is remembered between
submissions. Having it reset every time is annoying.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Ic681a3862e59d958fdebc5f9295d150f54d5a8c4