Write out full resolution floats in flatbuffer to JSON
It turns out that we were missing a 1--2 digits of precision
necessary to fully represent floats/doubles. In most cases where we
would care about bit-perfect

This moves all of our flatbuffer->JSON conversion to use
std::format/std::to_chars, which specifically guarantee that you will be
able to read the number back fully, while truncating numbers reasonably
so that e.g. 0.1 actually renders reasonably to the human eye.

In doing so, this also reduces some of the inconsistencies between the
two FlatbufferToJson methods.

This forces updates in a variety of tests that make use of this code and
which check for exactly identical serialized JSON values.

Change-Id: Idbf6a5614043ce4d6c02f3104991e30fdca23334
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>