Ignore node specific maps when no node is specified
In some contexts, we do a channel lookup without a node.  This is
especially true in LogReader when we are looking to register a mutation
callback for a message.  In this case, we were matching node specific
maps when we shouldn't be.  The end result is that subscribing to `/aos`
was resolving to `/pi2/aos` since that was the first remap found.

Change-Id: I428a01d24c7576ad9588b48ffbfb245d4bd98994
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>