Fix signs on wheels pointed backwards
When we go backwards, we want to flip the sign of the lateral force, and
want to also flip the direction of the contact patch piece of the self
aligning moment.

The easiest way to do this is to take the sin of the angle, rather than
doing a piecewise linear function of the angle to flip the sign.

Change-Id: I7b348a7c0b9449c28c01957aea7d4a16a52b863b
Signed-off-by: Austin Schuh <austin.linux@gmail.com>