Use casadi builtins to generate more optimal code
Casadi makes each function into a node on the execution graph.  When we
are doing long horizon optimization problems, we want the graph to reuse
as many functions as possible so the generated code ends up being
decently compact.  This keeps compilation time down to something
reasonable.

Turn on JIT compilation too while we are here.

Change-Id: I798186e52e77483bf616a064c81d23229d4c97c7
Signed-off-by: Austin Schuh <austin.linux@gmail.com>