Scouting app: fix action submission
In the Review screen, ending the teleop phase showed up as "9" since it was not caught in the switch case in entry.ng.html. Additionally, if
the scout switched between the endgame and review screen multiple times,
there would be multiple penalty actions since the penalty action created
when they went to the review page would not be deleted when they go back
to the endgame screen. This patch addresses these problems.

Signed-off-by: Emily Markova <emily.markova@gmail.com>
Change-Id: I7db2f93a74691e04a0d4290c19de927eabf8b802