scouting: Make the deployment work with new rules_pkg
The rules_pkg upgrade from a couple of months ago broke the scouting
deployment. The behaviour of the `include_runfiles` attribute changed.
Instead of flattening the runfiles tree, it now creates a dedicated
".runfiles" tree for a binary. This is great, but unexpected.

This patch fixes the code to work with the new directory layout.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I1757a42a5733579ed16bfd76f3436fb55d9a9e04