Add JAX cost, and test to verify correctness
Turns out there's a bug in the casadi too, so fix that while we are
here.  We were using the unrotated radius for computing the cost
function torque instead of the rotated radius.

Split the MPC code out into a separate file as well to make it easier to
pull into other things.

Change-Id: Iee6c9999fa8b6a91d6963af4edba5cf92a085e9b
Signed-off-by: Austin Schuh <austin.linux@gmail.com>