Switch to using tensorflow_probability
This is technically less accurate, but is a lot less to validate since
we are having trouble getting this to converge.

To try to adhere to the "unit gaussian" assumption that neural networks
like, make the physics account for the action limit instead of the actor
network.  We can undo this after things work if we find it doesn't make
it converge better.

Also, switch from talking about "cost" to "reward" to adhere to other
implementations closer too.

Change-Id: I1b8df0a72946020b3e3da5cf46b3787e38cb94c9
Signed-off-by: Austin Schuh <austin.linux@gmail.com>