Mark sandbox-breaking calibration test "manual"
This test went outside of the bazel sandbox to try to locate files on
the local filesystem; mark it manual to prevent it being run by default
until we can move the log dependency into a proper dependendency.

Also, make it so that we actually print out stdout/stderr from the
process that was failing to make error messages clearer.

Change-Id: Id0f95144e65d6c515ef93b8705e4a8f10a637aa5
Signed-off-by: James Kuszmaul <jabukuszmaul+collab@gmail.com>