Lower apriltag pixel border
I'm doing this now that we're using the 8 parameter model. The reason
why this pixel border was so large before is because we couldn't trust
the detections we get on the edge of the image. But now that we can
detect them better we should be okay to turn this down.

Change-Id: I2cd07df885a2ba03ae51799b4dd95ccd53084f33
Signed-off-by: Maxwell Henderson <mxwhenderson@gmail.com>