Add schema checker to AOS config
The aos_config flattener did not check we had a schema for each channel
so when on the robot, things like aos_dump crash. This is a very easy
thing to miss especially for new people(like I did) and the segfault
is very cryptic so this should help with that. Also, fix broken ones.

Change-Id: I11df08c5a9d4f1e180f4391b602c1fa1de4e518f
Signed-off-by: Justin Turcotte <jjturcot@gmail.com>