Add 2025 localizer
The goal of this is to give us robust localization on the field this
year, but because we haven't done full localization with swerve yet
we're implementing it using wpilib, weighted averaging, and a kalman
filter. More is explained in the comment in localizer.h

This also removes the x, y, and theta estimation in naive_estimator. It
no longer outputs x and y and the theta it uses for correction in the
lqr is taken from the localizer.

Furthermore this commit also edits autonomous to properly send out its
starting position when we press the user button.

Change-Id: I6e0734666b53b724d8d394413bea8920ba1f8103
Signed-off-by: Maxwell Henderson <mxwhenderson@gmail.com>