Fix imu_zeroer not un-faulting
We used to use FaultBehavior::kLatch but now we switch to
FaultBehavior::kTemporary and test that we fault in the middle of
getting a significantly offset zero but not at the end after we
re-zero.

Also, fault on recieving too many identical measurements, and add a
allow_duplicate_samples flag to toggle this behavior to use in tests
that use duplicate samples.

Signed-off-by: James (Peilun) Li <jamespeilunli@gmail.com>
Change-Id: I7272ed0549e2d2c3df728e2b2f0d797541827f06
Signed-off-by: James (Peilun) Li <jamespeilunli@gmail.com>